requirements.txt
google-generativeai
python-dotenv
langchain
scikit-learn
langchain-community
langchain_google_genai
flask